+++ _examples/1_SpectralDensity.py
@@ -11,20 +11,20 @@ sd = SpectralDensity.from_stream(stream, inv=inv)
 print(sd)
 
 # plot results
-base = "1_SpectralDensity"
-sd.plot(outfile=f'{base}_plot.png')
+file_base = "1_SpectralDensity"
+sd.plot()  #outfile=f'{file_base}_plot.png')
 
 # plot results, overlaid
-sd.plot(overlay=True, outfile=f'{base}_plot_overlay.png')
+sd.plot(overlay=True)  # , outfile=f'{file_base}_plot_overlay.png')
 
 # plot coherences
-sd.plot_coherences(display="full", outfile=f'{base}_coher_full.png')
+sd.plot_coherences(display="full")  #  , outfile=f'{file_base}_coher_full.png')
 
 # plot coherences, overlaid
-sd.plot_coherences(display="overlay", outfile=f'{base}_coher_overlay.png')
+sd.plot_coherences(display="overlay")  # , outfile=f'{base}_coher_overlay.png')
 
 # plot coherences, sparse
-sd.plot_coherences(display="sparse", outfile=f'{base}_coher_sparse.png')
+sd.plot_coherences(display="sparse")  # , outfile=f'{file_base}_coher_sparse.png')
 
-# plot coherences, minimal
-sd.plot_coherences(display="minimal", outfile=f'{base}_coher_minimal.png')
+# plot coherences, taking up minimal space
+sd.plot_coherences(display="minimal")  # , outfile=f'{file_base}_coher_minimal.png')